Move shared code into devtools_shared
