+++ packages/devtools_app/lib/src/shared/initializer.dart
@@ -6,13 +6,13 @@
 
 import 'dart:async';
 
+import 'package:devtools_shared/devtools_shared.dart';
 import 'package:flutter/material.dart';
 
 import '../analytics/analytics.dart' as ga;
 import '../analytics/constants.dart' as analytics_constants;
 import '../framework/framework_core.dart';
 import '../primitives/auto_dispose_mixin.dart';
-import '../primitives/url_utils.dart';
 import '../primitives/utils.dart';
 import 'common_widgets.dart';
 import 'globals.dart';
