+++ packages/devtools_shared/lib/devtools_shared.dart
@@ -9,3 +9,4 @@ export 'src/memory/event_sample.dart';
 export 'src/memory/heap_sample.dart';
 export 'src/memory/heap_space.dart';
 export 'src/memory/memory_json.dart';
+export 'src/service_utils.dart';
