+++ packages/devtools_shared/pubspec.yaml
@@ -14,3 +14,6 @@ dependencies:
   usage: ^4.0.0
   vm_service: ^8.1.0
   webkit_inspection_protocol: '>=0.5.0 <2.0.0'
+
+dev_dependencies:
+  test: ^1.0.0
